change from wrapper to `setUp` `tearDown` pattern

+++ tests/test_usersettings.py
@@ -8,46 +8,52 @@ sys.path.append(__file__ + "/../../src")
 from usersettings import UserSettings
 
 
-def preserve_settings(func):
-    def wrapper(*args, **kwargs):
+class UserSettingsTest(unittest.TestCase):
+    def setUp(self) -> None:
         original = os.path.abspath(__file__ + "/../../user_settings.json")
         backup = os.path.abspath(__file__ + "/../../user_settings.json.bak")
-        os.rename(original, backup)
-        try:
-            func(*args, **kwargs)
-        finally:
-            if os.path.exists(original):
-                os.remove(original)
-            os.rename(backup, original)
+        if os.path.exists(original) and not os.path.exists(backup):
+            os.rename(original, backup)
+        else:
+            os.remove(original)
 
-    return wrapper
+    def tearDown(self):
+        original = os.path.abspath(__file__ + "/../../user_settings.json")
+        backup = os.path.abspath(__file__ + "/../../user_settings.json.bak")
+        if os.path.exists(original):
+            os.remove(original)
+        os.rename(backup, original)
+        UserSettings._instance = None
 
-
-class UserSettingsTest(unittest.TestCase):
     def test_singleton(self) -> None:
         us1 = UserSettings()
+        self.assertTrue(us1.settings["App ID"] == "")
+
         us2 = UserSettings()
         self.assertEqual(us1, us2)
 
-    @preserve_settings
     def test_changed_settings_apply_to_all(self) -> None:
         us1 = UserSettings()
+        self.assertTrue(us1.settings["App ID"] == "")
+
         us2 = UserSettings()
         us1.settings["App ID"] = "123"
         self.assertEqual(us2.settings["App ID"], "123")
 
-    @preserve_settings
     def test_save_to_file(self) -> None:
         us = UserSettings()
+        self.assertTrue(us.settings["App ID"] == "")
+
         us.settings["App ID"] = "123"
         us.save_to_file()
         with open(os.path.abspath(__file__ + "/../../user_settings.json"), "r") as f:
             saved = json.load(f)
         self.assertEqual(saved["App ID"], "123")
 
-    @preserve_settings
     def test_load_from_file_overwrites_settings(self) -> None:
         us = UserSettings()
+        self.assertTrue(us.settings["App ID"] == "")
+
         us.settings["App ID"] = "123"
         us.save_to_file()
         self.assertEqual(us.settings["App ID"], "123")
@@ -56,9 +62,10 @@ class UserSettingsTest(unittest.TestCase):
         us.load_from_file()
         self.assertEqual(us.settings["App ID"], "123")
 
-    @preserve_settings
     def test_load_from_file_creates_file_if_not_exists(self) -> None:
         us = UserSettings()
+        self.assertTrue(us.settings["App ID"] == "")
+
         us.load_from_file()
         self.assertTrue(
             os.path.exists(os.path.abspath(__file__ + "/../../user_settings.json"))
